sparse-random: Explicitly cast UINT64_MAX to double to avoid warning
Clang complains correctly that UINT64_MAX is not representable exactly
in a double.  In this case it's fine, so cast explicitly to avoid the
warning.

sparse-random.c:215:30: error: implicit conversion from 'unsigned long' to 'double' changes value from 18446744073709551615 to 18446744073709551616 [-Werror,-Wimplicit-const-int-float-conversion]
    if (xrandom (&rs) <= P * UINT64_MAX)
                           ~ ^~~~~~~~~~

(cherry picked from commit 0122828edf860a38bb44b81dd537873cc2d7d6f7)

+++ plugins/sparse-random/sparse-random.c
@@ -205,7 +205,7 @@ sparse_random_get_ready (void)
      * (hole->data).
      */
     P = state ? P_dh : P_hd;
-    if (xrandom (&rs) <= P * UINT64_MAX)
+    if (xrandom (&rs) <= P * (double) UINT64_MAX)
       state ^= 1;
   }
